Move dict_almost_equal to test_io.py
