+++ topostats/io.py
@@ -33,48 +33,6 @@ CONFIG_DOCUMENTATION_REFERENCE = """# For more information on configuration and 
 # pylint: disable=too-many-lines
 
 
-def dict_almost_equal(dict1, dict2, abs_tol=1e-9):
-    """Recursively check if two dictionaries are almost equal with a given absolute tolerance.
-
-    Parameters
-    ----------
-    dict1: dict
-        First dictionary to compare.
-    dict2: dict
-        Second dictionary to compare.
-    abs_tol: float
-        Absolute tolerance to check for equality.
-
-    Returns
-    -------
-    bool
-        True if the dictionaries are almost equal, False otherwise.
-    """
-    if dict1.keys() != dict2.keys():
-        return False
-
-    LOGGER.info("Comparing dictionaries")
-
-    for key in dict1:
-        LOGGER.info(f"Comparing key {key}")
-        if isinstance(dict1[key], dict) and isinstance(dict2[key], dict):
-            if not dict_almost_equal(dict1[key], dict2[key], abs_tol=abs_tol):
-                return False
-        elif isinstance(dict1[key], np.ndarray) and isinstance(dict2[key], np.ndarray):
-            if not np.allclose(dict1[key], dict2[key], atol=abs_tol):
-                LOGGER.info(f"Key {key} type: {type(dict1[key])} not equal: {dict1[key]} != {dict2[key]}")
-                return False
-        elif isinstance(dict1[key], float) and isinstance(dict2[key], float):
-            if not np.isclose(dict1[key], dict2[key], atol=abs_tol):
-                LOGGER.info(f"Key {key} type: {type(dict1[key])} not equal: {dict1[key]} != {dict2[key]}")
-                return False
-        elif dict1[key] != dict2[key]:
-            LOGGER.info(f"Key {key} not equal: {dict1[key]} != {dict2[key]}")
-            return False
-
-    return True
-
-
 def read_yaml(filename: str | Path) -> dict:
     """Read a YAML file.
